Added skeleton code for DetourInterface and added debug ifdefs to RecastInterface debugging code.

+++ Wasteland_OGRE/Code/EWS.cpp
@@ -95,7 +95,7 @@ void EWSManager::Update(int newTime,bool isPlacing,const sPlayerData& playerData
 				//ammo in magazine
 				float magAmmoPcnt = static_cast<float>(_playerData.ammoInMag) / static_cast<float>(_playerData.magSize);
 				float height = magAmmoPcnt * 300;
-				Box(Ogre::Rect(300,400 - (height / 2),350,400),Ogre::ColourValue(.6f,.01f,.01f,1.0f));
+				Box(Ogre::Rect(300,static_cast<long>(400 - (height / 2)),350,400),Ogre::ColourValue(.6f,.01f,.01f,1.0f));
 				
 				//number of mags left
 				//write number on texture
